login bug correction if credential are not correct.

+++ frontend/index.php
@@ -21,25 +21,29 @@ $session = new Session();
 switch ($query_array['page']) {
     case '':
     case 'home':
-        $twigRenderer->home('');
+        $twigRenderer->home();
         break;
     case 'posts':
         $twigRenderer->posts();
         break;
     case 'post':
         $postId = $query_array['postid'];
-        $twigRenderer->post($postId, '');
+        $twigRenderer->post($postId);
         break;
     case 'login':
         $twigRenderer->login();
         break;
     case 'sendlogin':
         $loginForm = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-        $function->login($loginForm);
+        $res = $function->login($loginForm);
+        if($res === 'loginfail') {
+            $twigRenderer->login($res);
+        }
         break;
     case 'sendsignup':
         $signForm = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
         $function->signup($signForm);
+        $twigRenderer->login();
         break;
     case 'signup':
         $twigRenderer->signup();
@@ -49,48 +53,76 @@ switch ($query_array['page']) {
         $session->unsetSession();
         break;
     case 'admin':
-        $twigRenderer->adminPage();
+        if ($session->getSession('role') === 'admin') {
+            $twigRenderer->adminPage();
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'approvecomment':
-        $commentId = $query_array['commentid'];
-        $function->approveComment($commentId);
-        header('location: ?page=admin');
+        if ($session->getSession('role') === 'admin') {
+            $commentId = $query_array['commentid'];
+            $function->approveComment($commentId);
+            $twigRenderer->adminPage();
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'deletecomment':
-        $commentId = $query_array['commentid'];
-        $function->deleteComment($commentId);
-        header('location: ?page=admin');
+        if ($session->getSession('role') === 'admin') {
+            $commentId = $query_array['commentid'];
+            $function->deleteComment($commentId);
+            $twigRenderer->adminPage();
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'addcomment':
-        $comment = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-        $resCommentSent = $function->addComment($comment);
-        $twigRenderer->post($comment['postid'], $resCommentSent);
+        if ($session->getSession('role') === 'user' || $session->getSession('role') === 'admin') {
+            $comment = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+            $resCommentSent = $function->addComment($comment);
+            $twigRenderer->post($comment['postid'], $resCommentSent);
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'deletepost':
-        $postId = $query_array['postid'];
-        $function->deletePost($postId);
+        if ($session->getSession('role') === 'admin') {
+            $postId = $query_array['postid'];
+            $function->deletePost($postId);
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'updatepost':
-        $postId = $query_array['postid'];
-        $twigRenderer->updatePost($postId);
+        if ($session->getSession('role') === 'admin') {
+            $postId = $query_array['postid'];
+            $twigRenderer->updatePost($postId);
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'sendupdatedpost':
-        $postId = $query_array['postid'];
-        $form = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-        $function->updatePost($postId, $form);
+        if ($session->getSession('role') === 'admin') {
+            $postId = $query_array['postid'];
+            $form = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+            $function->updatePost($postId, $form);
+            $twigRenderer->updatePost($postId);
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'createpost':
-        $twigRenderer->createPost();
+        if ($session->getSession('role') === 'admin') {
+            $twigRenderer->createPost();
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'sendcreatepost':
-        $form = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-        $function->createPost($form);
-        $twigRenderer->posts();
+        if ($session->getSession('role') === 'admin') {
+            $form = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+            $function->createPost($form);
+            $twigRenderer->posts();
+        }
+        $twigRenderer->home('', 'true');
         break;
     case 'contact':
         $form = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
         $resSendEmail = $function->sendEmail($form);
-        $twigRenderer->home($resSendEmail);
+        $twigRenderer->home($resSendEmail, '');
         break;
     default:
         echo('page not found');
